requirements.txt
razorpay>=1.3.0
Flask>=2.0.0
mysql-connector-python>=8.0.26